aider: Added a button and input field for users to enter their own custom permutation.
# Aider chat conversation:

USER: I want to add a way of letting the user make their own custom permutations. what would be a good notation for that?

ASSISTANT: A common and intuitive way to represent custom permutations for a sliding puzzle would be to use a comma-separated list of numbers. This would allow the user to easily specify the order of the tiles.

For example, a user could input "1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,16" to create a puzzle where the 14 and 15 tiles are swapped.

To implement this, we would need to add an input field to the HTML where users can enter their custom permutation, and then modify the JavaScript to parse this input and use it to set up the puzzle.

Does this approach sound good to you?

USER: I like this. we could make this field under the "Goal" row. It could have a button "My Shuffle" and then a field that is prefilled with a permutation that the user can change. maybe instead of "16" it would be more intutitive to use "_" or "-".

ASSISTANT: Great! We can definitely use "_" or "-" to represent the empty tile. Let's use "_".

Here are the steps we need to take:

1. Add a new button and input field to the HTML.
2. Add an event listener to the new button in the JavaScript.
3. Write a function to parse the input and set up the puzzle.

Let's start with the HTML changes. We need to modify the `index.html` file.

```html
index.html
<<<<<<< HEAD
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
    </div>
=======
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
        <button id="my-shuffle-button">My Shuffle</button>
        <input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,_">
    </div>
>>>>>>> updated
```

Next, we need to add an event listener to the new button and write a function to parse the input and set up the puzzle. We need to modify the `script.js` file.