aider: Adjust the width of the input field in styles.css to make the permutation visible.
# Aider chat conversation:

USER: can you make the field with the permutation longer so that one can see all numbers?

ASSISTANT: Sure, we can adjust the width of the input field to accommodate the full permutation. We will need to modify the `styles.css` file to do this.

Could you please add the `styles.css` file to the chat so that I can propose the necessary changes?

USER: I added these *read-write* files: styles.css

ASSISTANT: Great! Now we can adjust the width of the input field. Here is the change we need to make in the `styles.css` file:

```css
styles.css
<<<<<<< HEAD
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }
=======
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }

    #my-shuffle-input {
        width: 400px; /* Adjust this value to change the width of the input field */
    }
>>>>>>> updated
```

This will make the input field wider so that the user can see the entire permutation. You can adjust the width as needed by changing the value of `400px`.