aider: Wrap the "My Shuffle" button and the input field in a div to visually group them together.
# Aider chat conversation:

USER: how could we change the layout so that it is clear that the "My Shuffle" button goes with the field after it?

ASSISTANT: To make it clear that the "My Shuffle" button goes with the field after it, we can make some changes to the HTML and CSS files.

In the HTML file, we can wrap the "My Shuffle" button and the input field in a div. This will group them together visually.

In the CSS file, we can add some styling to this new div to further emphasize that the button and the field are related. For example, we could add some padding around the div and a border around it.

Here are the changes:

**index.html**
```html
index.html
<<<<<<< HEAD
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
        <button id="my-shuffle-button">My Shuffle</button>
        <input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,_">
    </div>
=======
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
        <div class="my-shuffle-group">
            <button id="my-shuffle-button">My Shuffle</button>
            <input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,_">
        </div>
    </div>
>>>>>>> updated
```

**styles.css**
```css
styles.css
<<<<<<< HEAD
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }
=======
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }

    .my-shuffle-group {
        display: inline-block;
        padding: 10px;
        border: 2px solid #000;
    }
>>>>>>> updated
```

These changes will group the "My Shuffle" button and the input field together in a box, making it clear that they are related.