aider: Reverted changes and put "My Shuffle" button and input field in a new line.
# Aider chat conversation:

USER: do people use "shuffle" and "permutation" differently? should I prefer one term over the otehr?

ASSISTANT: In the context of puzzles and games, "shuffle" and "permutation" can often be used interchangeably. However, they do have slightly different connotations.

"Shuffle" is a term that is often used in everyday language. It suggests a random rearrangement of items, like shuffling a deck of cards. When you shuffle something, you don't know what the end result will be.

"Permutation", on the other hand, is a more mathematical term. It refers to all possible arrangements of a set of items. When you talk about permutations, you're often interested in the number of different ways you can arrange something.

In the context of your sliding puzzle game, either term could be appropriate. If you're talking about randomly rearranging the tiles, "shuffle" would be a good term to use. If you're talking about a specific arrangement of tiles, or the number of possible arrangements, "permutation" would be more appropriate.

In terms of which term to prefer, it depends on your audience. If you're talking to a general audience, "shuffle" might be more easily understood. If you're talking to a more technical or mathematical audience, "permutation" might be the better choice.

USER: can you revert your last change and instead put "My Shuffle" simply in a new line?

ASSISTANT: Sure, we can revert the changes and put the "My Shuffle" button and the input field in a new line. We will need to modify the `index.html` and `styles.css` files.

Here are the changes:

**index.html**
```html
index.html
<<<<<<< HEAD
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
        <div class="my-shuffle-group">
            <button id="my-shuffle-button">My Shuffle</button>
            <input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,_">
        </div>
    </div>
=======
    <div class="button-row">
        <button id="start-button">Goal</button>
        <button id="shuffle-button">Shuffle</button>
        <button id="puzzle14-button">14</button>
    </div>
    <div class="my-shuffle-group">
        <button id="my-shuffle-button">My Shuffle</button>
        <input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,13,15,14,_">
    </div>
>>>>>>> updated
```

**styles.css**
```css
styles.css
<<<<<<< HEAD
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }

    .my-shuffle-group {
        display: inline-block;
        padding: 10px;
        border: 2px solid #000;
    }
=======
    .button-row {
        display: block;
        height: 50px; /* or whatever height you want */
        text-align: left;
        margin-top: 100px; /* Increase this value to move the button further down */
    }

    .my-shuffle-group {
        display: block;
        padding: 10px;
    }
>>>>>>> updated
```

These changes will put the "My Shuffle" button and the input field in a new line.