aider: Set the size attribute of the input field to the length of the default string.
# Aider chat conversation:

ASSISTANT: Yes, you can achieve this by setting the `size` attribute of the input field to the length of the default string. This will make the width of the input field automatically adapt to the length of the default string.

Here is the necessary change to the `index.html` file:

```html
index.html
<<<<<<< HEAD
<input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,15,13,14,_">
=======
<input id="my-shuffle-input" type="text" value="1,2,3,4,5,6,7,8,9,10,11,12,15,13,14,_" size="35">
>>>>>>> updated
```

This change sets the `size` attribute to 35, which is the length of the default string "1,2,3,4,5,6,7,8,9,10,11,12,15,13,14,_". You may need to adjust this number to get the exact width you want, as the width per character can vary depending on the font and other styling.