aider: Remove width property from #my-shuffle-input selector.
# Aider chat conversation:

ASSISTANT: Yes, you're correct. The `width` property in CSS will override the `size` attribute of the input field. If you want the width of the input field to automatically adapt to the length of the default string, you can remove the `width` property from the CSS.

Here is the necessary change to the `styles.css` file:

```css
styles.css
<<<<<<< HEAD
#my-shuffle-input {
    width: 400px; /* Adjust this value to change the width of the input field */
}
=======
#my-shuffle-input {
    /* width property removed */
}
>>>>>>> updated
```

This change removes the `width` property from the `#my-shuffle-input` selector, allowing the `size` attribute in the HTML to determine the width of the input field.