aider: Fixed alignment and vertical space issues by resetting margins, padding, and line-height for button and div elements, and removing height rule from #button-row.
# Aider chat conversation:

USER: sth is still odd ... "My Shuffle" does not align with "Goal" and there is a vertical space between the two lines that should not be there

ASSISTANT: The alignment issue might be due to the default margins and padding applied by the browser to the `button` and `div` elements. We can fix this by setting the margins and padding to 0 for these elements in the CSS file.

The vertical space between the two lines might be due to the default line-height applied by the browser to the `div` elements. We can fix this by setting the line-height to 1 for these elements in the CSS file.

Here are the changes we need to make:

1. In the `styles.css` file, we need to add the following CSS rules to reset the margins, padding, and line-height for the `button` and `div` elements:

```css
styles.css
<<<<<<< HEAD
=======
button, div {
    margin: 0;
    padding: 0;
    line-height: 1;
}
>>>>>>> updated
```

2. In the `styles.css` file, we need to remove the `height: 50px;` rule from the `#button-row` selector. This rule is causing the extra vertical space between the two lines.

```css
styles.css
<<<<<<< HEAD
#button-row {
    display: block;
    height: 50px; /* or whatever height you want */
    text-align: left;
    margin-top: 100px; /* Increase this value to move the button further down */
}
=======
#button-row {
    display: block;
    text-align: left;
    margin-top: 100px; /* Increase this value to move the button further down */
}
>>>>>>> updated
```

Please confirm these changes so I can generate the final edit blocks.