aider: Fixed indentation issue for "My Shuffle" button and div elements.
# Aider chat conversation:

USER: "My Shuffle" is still indented

ASSISTANT: The indentation issue might be due to the default padding applied by the browser to the `button` elements. We can fix this by setting the padding to 0 for these elements in the CSS file.

Here are the changes we need to make:

1. In the `styles.css` file, we need to add the following CSS rule to reset the padding for the `button` elements:

```css
styles.css
<<<<<<< HEAD
button, div {
    margin: 0;
    padding: 0;
    line-height: 1;
}
=======
button {
    margin: 0;
    padding: 0;
}
div {
    margin: 0;
    line-height: 1;
}
>>>>>>> updated
```

Please confirm these changes so I can generate the final edit blocks.